Produce valid css when sprited images are missing.
If the last image in the list of images to sprite was missing, then the css
would have a spurious trailing comma. Now we only add the comma when
it is necessary.

Also improved the logging so that we just get an error message instead
of a stack trace when the file isn't found.